fix: correct spelling in API (AnnotationData, annotated_cols)
